Extract test module in a dedicated class; Add Javadoc
