+++ extensions/guice/src/main/java/me/dmmax/akka/spawn/protocol/guice/RegisterSpawnProtocolModule.java
@@ -3,16 +3,26 @@ package me.dmmax.akka.spawn.protocol.guice;
 import com.google.inject.AbstractModule;
 import com.google.inject.TypeLiteral;
 import com.google.inject.util.Types;
+import me.dmmax.akka.spawn.protocol.SpawnActor;
 import me.dmmax.akka.spawn.protocol.SpawnCommandConverter;
 import me.dmmax.akka.spawn.protocol.SpawnProtocol;
 
-public class RegisterSpawnProtocolModule<P> extends AbstractModule {
+/**
+ * Guice module to register spawn protocol for a specific type of actor.
+ *
+ * @param <T> - actor type
+ */
+public class RegisterSpawnProtocolModule<T> extends AbstractModule {
 
-  private final Class<P> type;
-  private final SpawnCommandConverter<P> spawnCommandConverter;
+  private final Class<T> actorType;
+  private final SpawnCommandConverter<T> spawnCommandConverter;
 
-  public RegisterSpawnProtocolModule(Class<P> type, SpawnCommandConverter<P> spawnCommandConverter) {
-    this.type = type;
+  /**
+   * @param actorType – actor type
+   * @param spawnCommandConverter – command converter which wraps {@link SpawnActor} to actor message type
+   */
+  public RegisterSpawnProtocolModule(Class<T> actorType, SpawnCommandConverter<T> spawnCommandConverter) {
+    this.actorType = actorType;
     this.spawnCommandConverter = spawnCommandConverter;
   }
 
@@ -20,14 +30,14 @@ public class RegisterSpawnProtocolModule<P> extends AbstractModule {
   @SuppressWarnings("unchecked")
   protected void configure() {
     // Register spawn command converter
-    TypeLiteral<SpawnCommandConverter<P>> spawnCommandConverterKey = (TypeLiteral<SpawnCommandConverter<P>>) TypeLiteral.get(
-        Types.newParameterizedType(SpawnCommandConverter.class, type));
+    TypeLiteral<SpawnCommandConverter<T>> spawnCommandConverterKey = (TypeLiteral<SpawnCommandConverter<T>>) TypeLiteral.get(
+        Types.newParameterizedType(SpawnCommandConverter.class, actorType));
     bind(spawnCommandConverterKey).toInstance(this.spawnCommandConverter);
     // Register spawn protocol
-    TypeLiteral<SpawnProtocol<P>> spawnProtocolKey = (TypeLiteral<SpawnProtocol<P>>) TypeLiteral.get(
-        Types.newParameterizedType(SpawnProtocol.class, type));
-    TypeLiteral<SpawnProtocolProvider<P>> spawnProtocolProviderKey = (TypeLiteral<SpawnProtocolProvider<P>>) TypeLiteral.get(
-        Types.newParameterizedType(SpawnProtocolProvider.class, type));
+    TypeLiteral<SpawnProtocol<T>> spawnProtocolKey = (TypeLiteral<SpawnProtocol<T>>) TypeLiteral.get(
+        Types.newParameterizedType(SpawnProtocol.class, actorType));
+    TypeLiteral<SpawnProtocolProvider<T>> spawnProtocolProviderKey = (TypeLiteral<SpawnProtocolProvider<T>>) TypeLiteral.get(
+        Types.newParameterizedType(SpawnProtocolProvider.class, actorType));
     bind(spawnProtocolKey).toProvider(spawnProtocolProviderKey);
   }
 }
